+++ OpenNos.Data/WearableInstanceDTO.cs
@@ -38,7 +38,7 @@ namespace OpenNos.Data
 
         public byte DarkElement { get; set; }
 
-        public byte DarkResistance { get; set; }
+        public short DarkResistance { get; set; }
 
         public short DefenceDodge { get; set; }
 
@@ -50,7 +50,7 @@ namespace OpenNos.Data
 
         public byte FireElement { get; set; }
 
-        public byte FireResistance { get; set; }
+        public short FireResistance { get; set; }
 
         public short HitRate { get; set; }
 
@@ -62,7 +62,7 @@ namespace OpenNos.Data
 
         public byte LightElement { get; set; }
 
-        public byte LightResistance { get; set; }
+        public short LightResistance { get; set; }
 
         public short MagicDefence { get; set; }
 
@@ -72,7 +72,7 @@ namespace OpenNos.Data
 
         public byte WaterElement { get; set; }
 
-        public byte WaterResistance { get; set; }
+        public short WaterResistance { get; set; }
 
         public long XP { get; set; }
